fixed endpoint spaces & added status progress checks using keys from skeletons
